data-collectors/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
pandas==2.1.3
numpy==1.26.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
schedule==1.2.0
fake-useragent==1.4.0
lxml==4.9.3
html5lib==1.1
scrapy==2.11.0
urllib3==2.1.0
python-dateutil==2.8.2
tenacity==8.2.3